# packages/vantage/src/vantage/service/openapi/v1.yaml
...
openapi: 3.1.0
info:
  title: Vantage API
  version: "1"
  description: read-tagged paths are guaranteed to leave stored data unchanged; write-tagged paths change stored data. Every rejection -- each 4xx and 5xx listed here, and the 404 or 405 for a path nothing serves or a method a path does not take -- has the Rejection body.
servers:
  - url: /api/v1
...
paths:
...
  /runs/{run_id}/results:
    get:
      operationId: list_results
      tags: [read]
      summary: Page through one run's results.
      parameters:
        - $ref: '#/components/parameters/run_id'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        "200": {description: A page of results., content: {application/json: {schema: {$ref: '#/components/schemas/ResultsResponse'}}}}
        "404": {description: No run with that id has been recorded., content: *rejection}
        "422": {description: "limit below 1 or offset out of range -- not a page; or run_id is not 32 lowercase hex characters.", content: *rejection}
...
  /config/sections:
    get:
      operationId: list_sections
      tags: [read]
      summary: List every defined test section.
      responses:
        "200": {description: Every stored section., content: {application/json: {schema: {$ref: '#/components/schemas/SectionListResponse'}}}}
        "500": {description: A stored section's value could not be parsed (unreadable_setting)., content: *rejection}
    post:
      operationId: upsert_section
      tags: [write]
      summary: Create or update a section's prefix. A section name may contain `/`, so it travels as a body field, never as a path segment.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SectionUpsertRequest'}}}}
      responses:
        "201": {description: Stored, a new section., content: {application/json: {schema: {$ref: '#/components/schemas/SectionResponse'}}}}
        "200": {description: Updated; the name was already stored., content: {application/json: {schema: {$ref: '#/components/schemas/SectionResponse'}}}}
        "422": {description: "An invalid or reserved name, an invalid prefix, the stored-section count bound, or a body that does not match SectionUpsertRequest.", content: *rejection}
    delete:
      operationId: delete_section
      tags: [write]
      summary: Delete a section by name.
      description: name is a query value, not a path segment, because a section name may contain `/`.
      parameters:
        - {name: name, in: query, required: true, schema: {type: string}}
      responses:
        "204": {description: Deleted.}
        "404": {description: No section with that name has been stored., content: *rejection}
        "422": {description: name missing., content: *rejection}
  /runs/{run_id}/sections:
    get:
      operationId: get_run_sections
      tags: [read]
      summary: One run's section summary -- every defined section, plus the always-present unassigned bucket.
      description: Section definitions are read fresh on every call, never cached; classification happens at read time.
      parameters:
        - $ref: '#/components/parameters/run_id'
      responses:
        "200": {description: The run's section summary., content: {application/json: {schema: {$ref: '#/components/schemas/RunSectionSummaryResponse'}}}}
        "404": {description: No run with that id has been recorded., content: *rejection}
        "422": {description: run_id is not 32 lowercase hex characters., content: *rejection}
        "500": {description: A stored section's value could not be parsed (unreadable_setting)., content: *rejection}
components:
  parameters:
    run_id: {name: run_id, in: path, required: true, schema: {type: string, pattern: '^[0-9a-f]{32}$'}}
...
    limit: {name: limit, in: query, required: false, description: "Page size. Values above 200 are clamped to 200; the response reports whether more items exist. A value below 1 is rejected with 422 -- it is not a page size.", schema: {type: integer, minimum: 1}}
    offset: {name: offset, in: query, required: false, schema: {type: integer, format: int64, minimum: 0}}
  schemas:
...
    FailureProjection:
      type: object
      required: [failure_type, failure_message, failure_message_truncated, failure_path, failure_lineno, skip_reason, xfail_reason]
      properties: {failure_type: {type: [string, "null"]}, failure_message: {type: [string, "null"]}, failure_message_truncated: {type: boolean}, failure_path: {type: [string, "null"]}, failure_lineno: {type: [integer, "null"]}, skip_reason: {type: [string, "null"]}, xfail_reason: {type: [string, "null"]}}
    ResultListItem:
      type: object
      required: [node_id, file_path, class_name, function_name, param_id, outcome, duration, started_at, finished_at, setup_outcome, call_outcome, teardown_outcome, setup_duration, call_duration, teardown_duration, worker_id, failure]
      properties: {node_id: {type: string}, file_path: {type: string}, class_name: {type: [string, "null"]}, function_name: {type: string}, param_id: {type: [string, "null"]}, outcome: {type: string, enum: [passed, failed, error, skipped, xfailed, xpassed]}, duration: {type: [number, "null"]}, started_at: {type: [string, "null"], format: date-time}, finished_at: {type: [string, "null"], format: date-time}, setup_outcome: {type: [string, "null"]}, call_outcome: {type: [string, "null"]}, teardown_outcome: {type: [string, "null"]}, setup_duration: {type: [number, "null"]}, call_duration: {type: [number, "null"]}, teardown_duration: {type: [number, "null"]}, worker_id: {type: [string, "null"]}, failure: {oneOf: [{$ref: '#/components/schemas/FailureProjection'}, {type: "null"}]}}
    ResultsResponse:
      type: object
      required: [items, has_more]
      properties: {items: {type: array, items: {$ref: '#/components/schemas/ResultListItem'}}, has_more: {type: boolean}}
    # The single-result endpoint's response -- every field a list response
    # bounds or excludes, unbounded, flat like `ResultReport`'s own wire
    # shape rather than nesting `failure`/`captured`.
...
    SectionUpsertRequest:
      type: object
      required: [name, prefix]
      properties: {name: {type: string}, prefix: {type: string}}
    SectionResponse:
      type: object
      required: [name, prefix]
      properties: {name: {type: string}, prefix: {type: string}}
    SectionListResponse:
      type: object
      required: [items]
      properties: {items: {type: array, items: {$ref: '#/components/schemas/SectionResponse'}}}
    SectionSummary:
      type: object
      required: [name, total, measured, passing, pass_percentage]
      properties: {name: {type: string}, total: {type: integer}, measured: {type: integer}, passing: {type: integer}, pass_percentage: {type: [number, "null"]}}
    RunSectionSummaryResponse:
      type: object
      required: [items, unassigned]
      properties: {items: {type: array, items: {$ref: '#/components/schemas/SectionSummary'}}, unassigned: {$ref: '#/components/schemas/SectionSummary'}}